// src/bases/base.interfaces.ts
export enum StatusCodes {
  SUCCESS = 'SUCCESS',
  FAILURE = 'FAILURE',
  BAD_REQUEST = 'BAD_REQUEST',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  FILE_NOT_ALLOW_MIME_TYPE = 'FILE_NOT_ALLOW_MIME_TYPE',
  FORMAT_ERROR = 'FORMAT_ERROR',
  EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
  EMAIL_INVALIDATE = 'EMAIL_INVALIDATE',
  DATE_OF_BIRD_INVALIDATE = 'DATE_OF_BIRD__INVALIDATE',
  DESCRIPTIION_INVALIDATE = 'DESCRIPTIION_INVALIDATE',
  URL_INVALIDATE = 'URL_INVALIDATE',
  FIELD_INVALIDATE = 'FIELD_INVALIDATE',
  INVALID_PASSWORD = 'INVALID_PASSWORD'
}

export interface ErrorResponse {
  data: any;
}